LPS-159861 Rename object-scripting-api
