+++ modules/apps/object/object-scripting-api/bnd.bnd
@@ -1,7 +1,7 @@
-Bundle-Name: Liferay Object Runtime API
-Bundle-SymbolicName: com.liferay.object.runtime.api
-Bundle-Version: 1.1.0
+Bundle-Name: Liferay Object Scripting API
+Bundle-SymbolicName: com.liferay.object.scripting.api
+Bundle-Version: 1.0.0
 Export-Package:\
-	com.liferay.object.runtime.scripting.exception,\
-	com.liferay.object.runtime.scripting.executor,\
-	com.liferay.object.runtime.scripting.validator
+	com.liferay.object.scripting.exception,\
+	com.liferay.object.scripting.executor,\
+	com.liferay.object.scripting.validator
