+++ modules/apps/object/object-scripting-api/src/main/resources/com/liferay/object/scripting/validator/packageinfo
